crud_api/urls.py
from django.urls import path
from rest_framework.urlpatterns import format_suffix_patterns
from crud_api import views

urlpatterns = [
    path('users/', views.UserList.as_view()),
    path('users/<int:pk>/', views.UserDetail.as_view()),
    path('expenses/', views.ExpenseList.as_view()),
    path('expenses/<int:pk>/', views.ExpenseDetail.as_view()),
    path('categories/', views.CategoryList.as_view()),
    path('categories/<int:pk>/', views.CategoryDetail.as_view())
]

urlpatterns = format_suffix_patterns(urlpatterns)
