requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.0.8
django-heroku==0.3.1
django-allauth==0.45.0
djangorestframework==3.11.0
gunicorn==20.1.0
cryptography==35.0.0
psycopg2-binary==2.8.6
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.3.0
cryptography==35.0.0
